refactor(#400): color picker to integrate more easily chart options
