+++ studio/src/components.d.ts
@@ -12,6 +12,9 @@ import {
   JSX,
 } from '@stencil/core';
 import {
+  InitStyleColor,
+} from './app/utils/editor/color.utils';
+import {
   Deck,
 } from './app/models/data/deck';
 import {
@@ -40,6 +43,18 @@ export namespace Components {
   }
   interface AppColor {
     'deckOrSlide': boolean;
+    'selectedElement': HTMLElement;
+  }
+  interface AppColorDeckSlide {
+    'applyToAllDeck': boolean;
+    'deckOrSlide': boolean;
+    'initCurrentColors': () => Promise<void>;
+    'moreColors': boolean;
+    'selectedElement': HTMLElement;
+  }
+  interface AppColorQrcode {
+    'initCurrentColors': () => Promise<void>;
+    'moreColors': boolean;
     'selectedElement': HTMLElement;
   }
   interface AppContact {}
@@ -208,6 +223,18 @@ declare global {
   const HTMLAppColorElement: {
     prototype: HTMLAppColorElement;
     new (): HTMLAppColorElement;
+  };
+
+  interface HTMLAppColorDeckSlideElement extends Components.AppColorDeckSlide, HTMLStencilElement {}
+  const HTMLAppColorDeckSlideElement: {
+    prototype: HTMLAppColorDeckSlideElement;
+    new (): HTMLAppColorDeckSlideElement;
+  };
+
+  interface HTMLAppColorQrcodeElement extends Components.AppColorQrcode, HTMLStencilElement {}
+  const HTMLAppColorQrcodeElement: {
+    prototype: HTMLAppColorQrcodeElement;
+    new (): HTMLAppColorQrcodeElement;
   };
 
   interface HTMLAppContactElement extends Components.AppContact, HTMLStencilElement {}
@@ -569,6 +596,8 @@ declare global {
     'app-avatar': HTMLAppAvatarElement;
     'app-code': HTMLAppCodeElement;
     'app-color': HTMLAppColorElement;
+    'app-color-deck-slide': HTMLAppColorDeckSlideElement;
+    'app-color-qrcode': HTMLAppColorQrcodeElement;
     'app-contact': HTMLAppContactElement;
     'app-create-slide': HTMLAppCreateSlideElement;
     'app-custom-data': HTMLAppCustomDataElement;
@@ -647,6 +676,18 @@ declare namespace LocalJSX {
   interface AppColor {
     'deckOrSlide'?: boolean;
     'onColorDidChange'?: (event: CustomEvent<boolean>) => void;
+    'selectedElement'?: HTMLElement;
+  }
+  interface AppColorDeckSlide {
+    'applyToAllDeck'?: boolean;
+    'deckOrSlide'?: boolean;
+    'moreColors'?: boolean;
+    'onColorChange'?: (event: CustomEvent<boolean>) => void;
+    'selectedElement'?: HTMLElement;
+  }
+  interface AppColorQrcode {
+    'moreColors'?: boolean;
+    'onColorChange'?: (event: CustomEvent<boolean>) => void;
     'selectedElement'?: HTMLElement;
   }
   interface AppContact {}
@@ -817,6 +858,8 @@ declare namespace LocalJSX {
     'app-avatar': AppAvatar;
     'app-code': AppCode;
     'app-color': AppColor;
+    'app-color-deck-slide': AppColorDeckSlide;
+    'app-color-qrcode': AppColorQrcode;
     'app-contact': AppContact;
     'app-create-slide': AppCreateSlide;
     'app-custom-data': AppCustomData;
@@ -890,6 +933,8 @@ declare module "@stencil/core" {
       'app-avatar': LocalJSX.AppAvatar & JSXBase.HTMLAttributes<HTMLAppAvatarElement>;
       'app-code': LocalJSX.AppCode & JSXBase.HTMLAttributes<HTMLAppCodeElement>;
       'app-color': LocalJSX.AppColor & JSXBase.HTMLAttributes<HTMLAppColorElement>;
+      'app-color-deck-slide': LocalJSX.AppColorDeckSlide & JSXBase.HTMLAttributes<HTMLAppColorDeckSlideElement>;
+      'app-color-qrcode': LocalJSX.AppColorQrcode & JSXBase.HTMLAttributes<HTMLAppColorQrcodeElement>;
       'app-contact': LocalJSX.AppContact & JSXBase.HTMLAttributes<HTMLAppContactElement>;
       'app-create-slide': LocalJSX.AppCreateSlide & JSXBase.HTMLAttributes<HTMLAppCreateSlideElement>;
       'app-custom-data': LocalJSX.AppCustomData & JSXBase.HTMLAttributes<HTMLAppCustomDataElement>;
